chore: Move the Attribute type to its own module
